unittests: Add first unittest, sysinfo_unittest
We'll start with a different convention than the one
used in autotest: We'll have a separate unittests/
directory with unittests, and those will load the
avocado libraries using simple import tricks (no
common.py).

This first version of the sysinfo unittest only
exercises the Loggable classes and their comparison
methods. More is to come (hopefully).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>