avocado.job: Implement Job and TestModuleRunner classes
A Job is a sequence of operations executed on a
machine. This patch implements the class, which
is a refactor of the test run function, and a
class that implements a convenience found in
python unittests: the main() method. This allows
test modules to be executable in stand alone
mode, without the need of the avocado test runner.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>