avocado.test: Make base test class inherit from unittest.TestCase
Compatibility among avocado and regular python unittests
is one of the goals we want to accomplish with making
test cases inherit from the base python unittest TestCase
class. Also, we get the benefit of interoperating with
other known python test frameworks, such as nose.

In order to do that, we make the Test() class to inherit
from the base TestCase class, then move the _run_test_instance
method to the test class (as it makes a lot more sense
there anyway), and turn it into the default runTest() method.

With this, avocado tests gain automatically all features
of the unittest.TestCase() class, namely, the assert
methods.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>