avocado.core.data_dir: Move avocado tmp location to /var/tmp
In newer systems, such as Fedora 19 onwards, the /tmp dir
is mounted on a tmpfs by default, and most users won't
ever disable this 'feature'. Some test suites were
not quite designed to run under tmpfs, hence they might
fail. Not to mention that sometimes the test suite might
be large, or generate large binary files so that tmpfs
won't be able to handle it.

So instead of storing test artifacts such as test suite
source code copies, and compiled test suites under /tmp,
let's use /var/tmp, that is not supposed to be wiped
through reboots, and usually is not mounted on tmpfs.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>