avocado.utils.params: Convert value type when specified in multiplex cfg
When we specify the type of a variable in the multiplex config:

    sleep_length = 1.0
    sleep_length_type = float

Now we'll convert automatically the string 1.0 retrieved
from the multiplex system to a float in a transparent way.
This will save code by making people to avoid to manually
convert values (if not specified, the value is considered
a string).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>