Reorganize unittests and functional tests
In order for people to run all avocado tests at once,
move and reorganize tests under a directory called
selftests. For someone to run all self tests in
avocado, one would have to execute:

    selftests/run selftests/all

This is a shortcut for developers. .travis.yml
was also updated to run tests based on the new
directory structure.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>