avocado.utils.process: Make functions to log stdout/err as it happens
Fixes bug #78

In order to make process.run() and friends to log stdout/err
as they are generated, create a BgSubProcess object that has
threads that poll those streams continuously, channeling the
outputs through the logging system, and also storing it in
the result attribute.

Maybe this implementation could use some more love to make
it fully thread safe, with locks. This is just a first
version to show it is possible in a cleaner fashion than
what is implemented in autotest currently.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>