avocado.job: Run avocado tests isolated in sub processes
In order to ensure avocado tests won't mess with the
main avocado process' environment, execute them in
forked subprocesses. If the test params have a
'timeout' attribute, use it as the timeout param
of the .join() multiprocess method, allowing the
subprocesses to be killed after timeout seconds if
they did not complete.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>