avocado.job: Always add JSON and Xunit outputs to job dir
As specified in the requirements meeting, always have Xunit and
JSON outputs written to the results dir. The basic idea is to
enable any output plugins specified in the cmd line, the Xunit
and JSON compulsorily, and in the end, if there are no output
plugins requested in the command line, enable human output.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>