avocado: Add the passfd module (pass file descriptors through UNIX domain sockets)
Python (< 3.3) lack support for the sendmsg/recvmsg libc functions, so this must
be achieved via an extension module that calls those functions.

A local (slightly modified) copy of passfd.c and passfd.py, from the python-passfd
project by Martin Ferrari implement most of this functionality.

The virt_passfd_setup.py file aids in the compilation and general use of this module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>