avocado: Add the machine module
Add the concept of machine classes in avocado. Machine
classes are the abstraction of a piece of hardware, that
offers you some functionalities, such as boot, reboot,
remote logins and the such.

The class specializes in two big categories:

* Bare Metal Machines: represent a physical machine, such
as a blade in a data center.
* Virtual Machines: represent a virtual machine, which
usually offers more possibilities than bare metal machines,
such as the ability to pause and take snapshots, as well
as migrate (move the state to another virtual machine).

The virtual machines can also be specialized, adding more
classes as we move forward. For example, QEMU virtual
machines have the Monitor class, that can interact with the
virtual machines in a richer way than a generic virtual machine.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>