Introduce test whiteboard
This introduces a simple file-like attribute of every single test,
named, you guessed it right, `whiteboard`.

This can be written multiple times during the test run, or a single
time at the very end of the test, as the test writter wishes.

It will be automatically closed (and flushed) when the test ends.

Signed-off-by: Cleber Rosa <crosa@redhat.com>