avocado: Use avocado.utils.path.init_dir in dir init operations
It saves quite a lot of code, since this is such a common
operation in avocado.

I'd like to have the reviewer to give a good thought of
whether this is a good solution or not (I think it is,
although API names and such could be better thought off).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>