avocado.utils.path: Add PathInspector class
We're planning to move away from the rule

1 test == 1 directory

In order to do that, we need to be able to tell
the difference between a dropin test (this can
be written in any language as long as it is an
executable file) and an avocado test (a python
file that contains a test class derived from
the base avocado test file).

Introduce in the path library a path inspector,
that has code to figure this out.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>