tests: Flatten tests directory
We want avocado tests to not require specific directories,
the user should be able to arrange them in the dir structure
they see fit. Also, if a test requires data, we suppose their
data dir is located in the same directory as the source file,
with the name:

[modulename].data

Re arrange the test dir to reflect this new structure.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>