avocado.core.data_dir: Changes in data dir format
Try as much as possible to turn timestamps into the
ISO format. In fact, iso requires hours and minutes
to be separated by a ':', but that makes URLs far
less convenient, requiring escape characters. So
let's replace it with a dot.

Also, put the short ID after the date, so date sort
can be properly done.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>