avocado: Omit tag on first instance, reproduce test dir structure
Instead of starting tags at 1 and compulsorily putting
them on tests (even if it is the 1st execution of the test),
start with 0 and omit it if it is only the first test.

Also, reproduce the directory structure of tests in the
logs directory. This means that a test execution of

mydir/firstlevel.py
mydir/category/secondlevel.py

Would translate in the final directory results as:

mydir/firstlevel.py/
mydir/category/secondlevel.py/

Preserving the directory structure makes things more
consistent and intuitive, while preserving a unique
test ID.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>