avocado: Fix install and execution under virtual environment
The use of virtual environments is a very useful feature
for developers and helps us to sanity check code instalability.
However, we were not able to use it because the data file
install has to be different for virt env and for regular
system wide install (RPMs).

This patch makes avocado aware of the execution mode,
and sets the install process so that it works differently
(and fine) for both normal install and virtenv install.

Additionally, this fixes a bug I introduced on installing
test files, that broke the avocado rpms.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>