avocado.job: Send SIGKILL on Ctrl+C to all subprocesses
Turns out some subprocesses might misbehave (see tests
like trinity - run on a VM), so it's not enough to send
the interrupt signal. For a clean stop output wise,
SIGKILL is necessary.

As a side effect, the previous Ctrl+C can be resurrected,
this time renamed to _interrupt, since we're not really
simulating Ctrl+C.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>