tests/sleeptenmin.py: implement a test specific progress check
Which doesn't hold any logic, but serves as an example.

To watch the intended behavior in action, I recommend running the
"one_hundred_cycles" variant of this tests.

The test runner will show that the test is running (yellow throbber
spinning), and once each six seconds (+1% into the test run) a green
throbber will be shown.

Signed-off-by: Cleber Rosa <crosa@redhat.com>