avocado.job: Change Ctrl+C handling
Instead of sending SIGKILL to child processes right away,
ignore Ctrl+C for a time window of 2 seconds, explaining
to the user that, and that after the ignore window,
new Ctrl+C will send SIGKILL to the children and terminate
the process.

The functional test was removed, and will be re-added
once I figure out how to properly test the interrupt
behavior.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>