avocado: Change framework to use the new multiplexer
That change has to be made all at once for the sake of
bisectability. Update avocado.job and the multiplexer
plugin with the new multiplexing library code.

Also, update functional tests, the avocado sample tests
with the new YAML based config files, and the requirements.txt
file with the new prominent dependency of avocado, PyYAML.

Signed-off-by: Ruda Moura <rmoura@redhat.com>