avocado.core.job_id: Ensure the Job ID is a SHA1
Return the SHA1 of the random data, rather than trusting
that the random number is large enough to fill the 40
character requirement.

This fixes a bug where on occasions the job ID had less
than 40 characters on Travis (later verified to happen
in our machines as well, we just didn't know).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>