avocado: Move tests to example/tests
We have tests/ and selftests/ in avocado (the test runner).
The former includes examples of tests that make use of the
avocado API, and are also used to test avocado via CI.

This is confusing, as it doesn't make it clear to users that
tests are not part of the avocado framework. Since we have a
directory called examples with some other files (multiplexer,
plugins), move the tests to there (and keep using them in the CI).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>