journal plugin: put the journal file in the right directory
Here right directory means the top log job result directory,
along with the job id file and the job.log file, etc.

Looks like there was a change in the meaning of `job.logdir`
and because of that, the journal plugin was saving the file
one directory up the tree.

Also, we needed the job logdir as part of the test state, just
like we need the job unique id.

Signed-off-by: Cleber Rosa <crosa@redhat.com>