avocado.multiplexer: filters accepts strings or lists of strings.
For the filter_only and filter_out functions for filtering,
always accept a string or a list of string for for the parameters.

So this is possible to parse a YAML file like:

foo:
    filter-only: /root/env/prod

or

foo:
    filter-only: [/root/env/prod]

The previous behavior was treating the string as sequence,
so the filters were using the chars as input values.

Signed-off-by: Ruda Moura <rmoura@redhat.com>