avocado.plugins.vm: Display Job ID properly in the CLI
Set a Job ID in the host, and force the avocado command
run using it inside the VM. With this, we get to display
that information to the user.

Also, right now, as we parse and display results all at once,
displaying the job log doesn't matter much. Therefore,
let's keep the output symmetrical to the one obtained
running avocado in the host.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>