avocado.core.tree: Introduce function apply_filters().
Introduce function apply_filters(), which will prune
the tree that we use to multiplex and create variants,
by using rules for filter only and filter out, passed by
the command line (aka global filters).

It works for the test runner and for the multiplex plugin,
or anything that uses the TreeNode data structure with filters.

Signed-off-by: Rudá Moura <rmoura@redhat.com>