avocado.multiplexer: Ignore root node if it is leaf.
When getting leaves, if the node is root and alone, ignore it.
This situation is possible if filtering just preserves the root node,
so the root node will be leaf and returned.

That makes things a bit weird to get the variants,
because it will become one. So, we are going to ignore it.

Signed-off-by: Rudá Moura <rmoura@redhat.com>