avocado.utils.process: Add record_stream_files param to APIs
Add the new record_stream_files to the process APIs, which
basically means, 'log stdout and stderr of that subprocess
to the test stdout.actual and stderr.actual files'. We
thought about making that transparent, but frequently people
will want to execute programs inside tests that are part
of the setup of those tests, for example, make commands and
the such.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>