examples/tests/synctest.py: Add record_stream_files arg to synctest call
Add the record_stream_files argument to the synctest
command execution, to showcase how to use it. The idea
is that the full output of the synctest command goes to
the 'stdout.actual' and 'stderr.actual' files, excluding
all previous 'make' commands used to build the test suite.

With this, we pave the way for a feature that people can
use to record the outputs of commands, and later compare
with the reference output automatically.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>