avocado: Change param record_stream_files to allow_output_check
After review, we settled in 'allow_output_check' rather than
'record_stream_files', since we are allowing some of the output
streams of the program to be checked. Also, in order to make
the feature easier to use, turn it on by default (set
allow_output_check=True by default).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>