avocado: Remove --loglevel and add --job-log-level to 'run' subcommand
It was an option that didn't work properly, and made
more sense to the run subcommand anyway. Now, you can
provide one of the valid logging levels in the python
logging level translated to a lowercase string
('debug', 'info', 'warning', 'error', 'critical') and
that will be the level used for the job log in both
disk and --show-job-log.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>