avocado.utils.remote: Changes to the .run() method
Make the run() method to return a CmdResult object,
to make it symmetrical to the process.run() result.
Also, turn on logging of outputs for commands executed
on remote connections.

The reason why we want to do that is to increase the
debugging output we have for commands executed over
the fabric SSH connections. This way we can obtain
better, richer outputs in test logs that use our
fabric wrappers.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>