avocado.utils.process: add support for environment variables.
The main SubProcess class now supports an argument (env) to receive
extra environment variables to the subprocess execution.

Note that it's extra in the sense that it includes the
current environment variables, so the previous environment is
not destroyed.

The utility functions now accepts this extra environment variables.
It's more robust to pass variables as a dictionary than
using the shell feature of shell variables.

Signed-off-by: Rudá Moura <rmoura@redhat.com>