GDB: introduce a "ready to use" check when instantiating a GDBServer
It turns out that, in some conditions, such as when immediatelly trying
to communicate with a gdbserver process after it is executed, results in
failure.

The reason seems to be that gdbserver itself takes "some" time between
being executed and binding to the TCP port.

While prototyping GDB remote client in pure Python, this kind of race
condition was found.

Now, by default, there will be a 2 seconds timeout and a explicit check
for gdbserver's "ready to use" status.

Changes from v1:
 * added documentation for the __init__ method

Signed-off-by: Cleber Rosa <crosa@redhat.com>