GDB: use gdbserver --server-sdtderr option if available
A new command line option is being added to upstream GDB. This new
option allows gdbserver's own stderr (but not its inferior process
stderr) to be redirected to another file.

If that feature is available on the about to be used gdbserver, then
let's make use of it. This redirects its stderr to /dev/null, because
avocado itself has no real use for gdbserver's own stderr output.

Signed-off-by: Cleber Rosa <crosa@redhat.com>