Avocado: Add htmlresult plugin
Add a plugin to write a job summary HTML page.
The plugin takes results and creates a JSON context, that
is used by the pystache library, to render an HTML page
using a mustache template.

This commit contains the base plugin code, and the resource
files used by the plugin (bootstrap for some of the js, base
css, datatables for the test results table functionality).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>