avocado.multiplexer: Construct the tree directly from yaml file
Instead of creating OrderedDicts and then converting it into
tree structure, this modifies the `mapping` constructor to create
the tree structure directly. This way we only walk through the yaml
file once.

Also remove the `tree_unittest.TestReadYAML` selftest as it lost it's
meaning. Valures are checked in the TestTreeNode selftest.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>