examples/tests/abort: Add a timeout to avoid aborttest taking too long
It is likely that a race condition in avocado makes the abort
test to take a long while executing (24 hours), but that
was only identified in the Travis CI environment. Instead
of taking that long, we should end the abort test quickly,
after a test timeout. This will allow us to continue to
see the failure, only that the rest of the tests will be
executed, and no penalties to Travis due to inefficient
builds.

Of course, we still need to find the actual bug and fix it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>