avocado.job: Fix abort code problem
On early test load process, we try to get the early
test state. In Travis VM environment, the call to
get the early state sometimes never completes.

Let's put a timeout to this call, so it doesn't
blocks forever. In case the message queue is empty,
fill the test properties inside the avocado process
(but be careful). This way we get the information
we're missing.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>