examples.wrappers: Turn ltrace.sh smarter.
Use 'set -e' to exit in the first error problem, for example, when
the ltrace command is not installed.

The script now looks inside the log to probe for exit code, after
that, it tries to look if the process has been killed by a signal.
This order is different from the previous implementation.

The regular expression to probe for exit code and signals are improved.

Signed-off-by: Rudá Moura <rmoura@redhat.com>