avocado.runner: Set test's queue when running the test.
Set test queue communication in the moment we're going to execute test.
So that, the test loader doesn't requires a queue when discovering the
tests.

Also the argument for `run_suite` is now a test suite: a list of
test factories and not a params_list (a list of dict).

Signed-off-by: Rudá Moura <rmoura@redhat.com>