avocado: Remove the concept of MissingTest
The idea of a test that doesn't exist was introduced to
show people that the absolute path provided by them didn't
exist (maybe they commited a typo), but it turns out this
is a bad idea in a world where we scan directories for tests.

Let's simply scan the paths specified, and if no tests
are present, tell that to the user. Ignore any tests that
do not exist.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>