avocado.loader: Better heuristics to detect avocado test classes
When inspecting a possible test file, instead of using the
poor correspondence class name -> file name, find the class
that is a subclass of avocado.test.Test. Much more reliable
and flexible.

Next commit updates the documentation. Later on, we're going
to introduce support for avocado to run any python unittest,
but for now, let's stick only with the avocado test classes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>