Distro plugin: add a new distro plugin and basic command
At this stage, the distro plugin is very simplistic and prints the
detected distribution of the current machine. One could algue that
this is rather useless, but, with the component isolation work this
command is going to be helpful for troubleshooting purposes.

Also, this plugin is a placeholder where other functionality will
live, including (but not limited to):

 * the ability to create "distro definition" files, akin to
   autotest's autotest-distro-def-create.
 * the ability to view "distro definition" files, akin to
   autotest's autotest-distro-def-view.
 * the ability to download "distro definition" files from a
   repository
 * (maybe) the ability to push a distro definition file to
   avocado-server. This may be delegated to a full blown REST
   client though.

Signed-off-by: Cleber Rosa <crosa@redhat.com>