avocado: Improve avocado test load heuristics
When avocado analyzes a path that looks like a python module,
first it'll try to import the module, then look for an appropriate
test class inside this module. Improve the test load heuristics
by looking for classes that inherit from avocado.test.Test, and
not the rudimentary detection, inspired by autotest.

If the module import went fine, but we simply can't find an
avocado test inside of it, if the file is executable we'll
try to execute it as a simple test. If it's not executable,
it'll be considered as not a test (same count as not_found tests).
If we can't import the module due to a test bug, once again we'll
check if the file is executable - if executable, we'll execute it
as a simple test, if not, we'll consider it a buggy test.

Unfortunately there's no way to tell apart a buggy avocado test
and a random py module with a bug, so both will be treated the same.
Adjust code to reflect this new status.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>