avocado: Improve avocado test load heuristics
When avocado analyzes a path that looks like a python module,
first it'll try to import the module, then look for an appropriate
test class inside this module. Improve the test load heuristics
by looking for classes that inherit from avocado.test.Test, and
not the rudimentary detection, inspired by autotest.

If the module import went fine, but we simply can't find an
avocado test inside of it, classify it as NOT_A_TEST status
(same count as not_found tests), and if we can't import the
module due to a buggy module, we'll classify it as a BuggyTest.

Unfortunately there's no way to tell apart a buggy avocado test
and a buggy, random py module, so both will be treated the same.
Adjust code to reflect this new status.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>