avocado: Add another branch in test detection heuristics
It occurred to us that we can tell apart a buggy avocado
test from a random buggy module since all avocado instrumented
tests will have an avocado import. Therefore, let's do a regexp
search on the module to see if we find an avocado import header.

Of course, this still has room for mistakes, but this is a
best effort enterprise anyway. Update the diagram with the
extra branch in the test resolution.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>