avocado.runner: Improve test load failures debugability
The moment where we load the test in avocado is critical,
since it happens on a subprocess, a failure there is
opaque to the user, and the test early state never
gets populated, which makes avocado hang forever.

Handle this case better, and print an exception right
away to the user.

Changes from v1:
 * Improved error message, per Cleber's comments.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>