avocado.loader: Selectively ignore non tests on the test runner
Introduce a simple rule to whether show tests or not depending
on user's input: If it's a path to a file, then the user wants
avocado to explicitly run it, if it's a path to a directory,
then the user is interested only on actual tests. So treat
both entries differently, omitting non tests when the entry
is a dir, and showing non tests when the entry is an explicit
file.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>