Add runtime reference to the currently running job
We already have a reference in the runtime module to the currently
running test, but, even though we could reference the job from the
currently running test, the job may still be running after all tests
finish.

So, let's add an explicit reference to the currently running job.

Notice: IMHO, there's no need to define when a job finishes running
and clear it from the runtime module. It can be argued that another
job would only replace the current one if/when its run() method
would be called.

Signed-off-by: Cleber Rosa <crosa@redhat.com>